JS-12 Fixed other inconsistencies in transaction factory create methods

+++ sdk-vertx-client/src/test/java/io/nem/sdk/infrastructure/vertx/TransactionModelToJsonTest.java
@@ -396,8 +396,8 @@ class TransactionModelToJsonTest {
         MultisigAccountModificationTransaction transaction =
             MultisigAccountModificationTransactionFactory
                 .create(NetworkType.MIJIN_TEST,
-                    2,
-                    1,
+                    (byte) 2,
+                    (byte) 1,
                     Arrays.asList(multisigCosignatoryModification))
                 .build();
 
@@ -427,8 +427,8 @@ class TransactionModelToJsonTest {
         MultisigAccountModificationTransaction multisigAccountModificationTransaction =
             MultisigAccountModificationTransactionFactory
                 .create(NetworkType.MIJIN_TEST,
-                    2,
-                    1,
+                    (byte) 2,
+                    (byte) 1,
                     Arrays.asList(multisigCosignatoryModification))
                 .build();
 
@@ -464,8 +464,8 @@ class TransactionModelToJsonTest {
         MultisigAccountModificationTransaction multisigAccountModificationTransaction =
             MultisigAccountModificationTransactionFactory
                 .create(NetworkType.MIJIN_TEST,
-                    2,
-                    1,
+                    (byte) 2,
+                    (byte) 1,
                     Arrays.asList(multisigCosignatoryModification))
                 .build();
